RTCC: Use a simple system for the coordinate system epoch; a few preparations for switching to the correct RTCC coordinate system

+++ Orbitersdk/samples/ProjectApollo/src_rtccmfd/ARCore.h
@@ -155,7 +155,7 @@ public:
 	void GenerateAGCEphemeris();
 	int agcCelBody_RH(CELBODY *Cel, double mjd, int Flags, VECTOR3 *Pos = NULL, VECTOR3 *Vel = NULL);
 	int agcCelBody_LH(CELBODY *Cel, double mjd, int Flags, VECTOR3 *Pos = NULL, VECTOR3 *Vel = NULL);
-	void AGCEphemeris(double T0, double Epoch, double TEphem0);
+	void AGCEphemeris(double T0, int Epoch, double TEphem0);
 	void AGCCorrectionVectors(double mjd_launch, double t_land, int mission, bool isCMC);
 	void GenerateAGCCorrectionVectors();
 
@@ -372,7 +372,7 @@ public:
 
 	//AGC EPHEMERIS
 	int AGCEphemOption;	//0 = AGC ephemerides, 1 = AGC precession/nutation/libration correction vectors
-	double AGCEphemBRCSEpoch;
+	int AGCEphemBRCSEpoch;
 	double AGCEphemTEphemZero;
 	double AGCEphemTIMEM0;
 	double AGCEphemTEPHEM;
